ADD, EDIT
ADD : PanelToggleControl, DebugButton
EDIT : HappeningOccur
디버그용 이벤트 출력 버튼 만들고, HappeningOccur 클래스 정상작동하도록 수정
