+++ Assets/01. Scenes/Test Scenes/MainScreen.unity
@@ -269,7 +269,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 077c5384acb1fc54faf9c7b62925303e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  targetScene: Ui
+  targetScene: Ui(TextOutputTest)
 --- !u!1 &816329315
 GameObject:
   m_ObjectHideFlags: 0
@@ -694,6 +694,57 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 0}
   m_Pivot: {x: 0, y: 0}
+--- !u!1 &1412845527
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1412845529}
+  - component: {fileID: 1412845528}
+  m_Layer: 0
+  m_Name: HappeningUtils
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1412845528
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1412845527}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e6834765ebecaab49b885881e6342f76, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  startDate: 0/3/1
+  endDate: 1/7/31
+  summerVactionStart: 0/8/14
+  summerVactionEnd: 0/8/31
+  winterVactionStart: 0/1/6
+  winterVactionEnd: 0/1/31
+  springVactionStart: 0/2/8
+  springVactionEnd: 0/2/28
+--- !u!4 &1412845529
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1412845527}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 121.02403, y: 666.61426, z: -68.458244}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1602462619
 GameObject:
   m_ObjectHideFlags: 0
