+++ Assets/02. Scripts/CircleZero.cs
@@ -7,7 +7,7 @@ public class CircleZero : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        
+        // HappeningUtils.happeningUtils.ReadData(); 이렇게 쓰면 될듯?
     }
 
     // Update is called once per frame
