+++ ProjectSettings/EditorBuildSettings.asset
@@ -14,4 +14,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/01. Scenes/Test Scenes/Ui.unity
     guid: 95ce769b10c587d44990fbabfa776cdb
+  - enabled: 1
+    path: Assets/01. Scenes/Test Scenes/Ui(TextOutputTest).unity
+    guid: 69c789aab615ea049a7ae215d9dfda82
   m_configObjects: {}
