+++ UserSettings/EditorUserSettings.asset
@@ -9,25 +9,25 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
       flags: 0
     RecentlyUsedScenePath-1:
-      value: 224247031146460b5c404c11132651150558352b3f3c23226d1a1e36ece57a2decee22f0
-      flags: 0
-    RecentlyUsedScenePath-2:
-      value: 224247031146460b5c404c111326511505582b2b21382a351e2a183de7ae2136ebf32f
-      flags: 0
-    RecentlyUsedScenePath-3:
       value: 224247031146460b5c404c111326511505582c2f3f3c66032e2c1336f1af1e31dbe239e7a92f31352d1b
       flags: 0
-    RecentlyUsedScenePath-4:
+    RecentlyUsedScenePath-2:
       value: 224247031146460b5c404c111326511505582c2f3f3c66032e2c1336f1af1731f0e43aecfd3f2d330a01ea3201701431fb1e10
       flags: 0
-    RecentlyUsedScenePath-5:
-      value: 224247031146460b5c404c111326511505582c2f3f3c66032e2c1336f1af1939ebe905eaf53f3a327717e1351027
-      flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-3:
       value: 224247031146460b5c404c111326511505582c2f3f3c66032e2c1336f1af0131acf238e0f323
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-4:
+      value: 224247031146460b5c404c11132651150558352b3f3c23226d1a1e36ece57a2decee22f0
+      flags: 0
+    RecentlyUsedScenePath-5:
+      value: 224247031146460b5c404c111326511505582b2b21382a351e2a183de7ae2136ebf32f
+      flags: 0
+    RecentlyUsedScenePath-6:
       value: 224247031146460b5c404c111326511505582c2f3f3c66032e2c1336f1af0131aad333f1f3152a282917fb08012d1576bc1f0702e212
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 224247031146460b5c404c111326511505582c2f3f3c66032e2c1336f1af1939ebe905eaf53f3a327717e1351027
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
